Extract app to a separated file.
- The goal is to make app runtime as weak pure function.
- Assert services object as `createApp` function input.
- Regsiter services on `req.services` on every request.
